Quick API file/bundle full lifecycle test (#154)
* API file/bundle lifecycle test

Uploads files, creates a bundle, retrieves bundle, retrieves files.

The motivation for this test was that on 6/23 I found a bug when I went through the complete lifecycle of file upload, bundle upload, bundle download and file download. It wasn't until we exercised the complete lifecycle that we realized the downloaded files didn't refer to the correct bundle. It appears we don't have another test that exercised the complete lifecycle, and this is something we can test without needing to integrate with other systems, so we should test it.

While test_bundle and test_file should exhaustively test all the attributes returned from individual calls, this test tests the coordination between the bundles and files.