Exclude indexed files that cannot be parsed as JSON
If a bundle file is marked to be indexed, and has content type
application/json yet cannot be parsed as json, a warning is logged
by the index Lambda and that file is excluded from the index document
for that bundle.

Note: This commit adds test fixture data and requires populate.py
to be run for tests to pass.