Enable use of populate.py with moto S3 mock
Changed cloud_uploader.py support for S3 to use boto3 directly
rather than running AWS CLI commands. This enables populate.py to be
used in unit tests with the moto S3 mock. It also runs much faster
than it did before.

Populate.py was modified slightly to enable convenient programmatic
use, in addition to CLI use.

This was tested both in the context of moto unit tests as well as
with actual S3 through the CLI.

The corresponding changes have not been made for the cloud_uploader.py
GS support, in part because I am not aware of a good mock being available
for GS. However, it may be worth the conversion away from using CLI
commands for the performance improvement alone.