Fixes for post commit review comments for indexer-related code and tests (#245)
* Changed indexer test to use common/core test data

The indexer test had setup its own sample data before the
core storage operations were available. Now that the storage
operations and common test fixture data and test methods are
available, use those instead. Test methods were extracted
from test_api.py and made commonly available in infra.py.
Because the common test infrastructure uses S3 directly,
the use of the moto S3 mock framework was removed.
Also changed validation of the index document so the file
that represented the expected result is no longer needed
and was removed.

* Improve test_indexer.py use of Elasticsearch client.

Reduced the number of Elasticsearch client connections
to one over the life of the test.

* Exclude indexed files with incorrect content type

If a file in a bundle is marked to be indexed in the bundle manifest
yet the file does not have content type application/json, then the
indexer Lambda process will log a warning and ignore/exclude the file
from the index document.

Also updated index.py code to use core constants for BundleMetadata and BundleFileMetadata

* Exclude indexed files that cannot be parsed as JSON

If a bundle file is marked to be indexed, and has content type
application/json yet cannot be parsed as json, a warning is logged
by the index Lambda and that file is excluded from the index document
for that bundle.

Note: This commit adds test fixture data and requires populate.py
to be run for tests to pass.