Handle parameter mismatches and return 400.
Right now, parameter mismatches raise a generic TypeError.  We can't catch that because it could be an internal TypeError, and not necessarily tied to the parameters users send us.

Instead, we attempt to bind the functions to the args, and if that fails, then we know it's the arguments a user sends us.  If it passes, and some *other* TypeError is raised, we still send that to the generic exception handler.

Add test for this case.

Addresses issue #297