Fix intermittent test failure related to file blob removal (#329)
The indexer test test_indexed_file_access_error formerly removed
the a file blob associated with a test bundle to verify the indexer
handles this case as expected. However, the file blob was for a commonly used
test file, and intermittent test failures were occurring when run using a shared test bucket.
The test was changed to instead create a bundle manifest entry such that the file blob
reference is invalid. The code and test semantics are the same.