Make Elasticsearch indexes replica specific
Change Elasticsearch index names to be replica specific to enable replica
independent indexing, search, subscription/notification.
This commit is focused on replica specific indexing, with changes
to search and subscription for all tests to pass with replica specific indexes.
Additional changes are required to make the search route replica aware,
and additional GCP-specific testing is required for subscriptions,
which will be done as additional PRs.