Parallel copy for S3. (#421)
The supervisor task has the responsibility of setting up the multipart copy, starting up the workers, waiting for the part copies to complete, and then to join the parts into a final blob.

The worker task is given a range of parts to copy, and executes them with a ThreadPoolExecutor.

Adds tests to validate functionality.