Send a content-length header to prevent a deadlock. (#508)
If we don't send a content-length, the python requests library will sit there and continue to try to read until the network connection is severed.  This is happening because (after some of my other changes), there's a library that's polluting the http.server globals (I haven't tracked this down) and enabling HTTP/1.1, which supports resumes.