revert unnecessary changes
I did a poor job of reviewing #509.  I reviewed it for logical correctness rather than whether it makes sense as a whole.  The only changes we need are the changes to the static test data.  We don't need to tag the size data as metadata since that's freely available from the cloud provider.