spin-loop for PUT /bundles on missing file
We've observed that there are race conditions on S3 where a `PUT /files` has completed but the metadata files are not available on the subsequent `PUT /bundles` request.  We added logic to make it easier for the client to identify this situation (and retry) in #566, but it is desirable to attempt the retry on the server as well.

Connects to #464