Unit tests for multiple schema version indexing and search
Test with version 2 (unversioned) data bundles, in addition to current (version 3)
data bundles. Verify correct aliases and indices are crated and that both version-specific
and multi-version searches work as expected.

The populate.py script must be run to update the text fixture buckets for these tests
to pass.

Tests with version 4 and later data will be added when the data is available.