[WIP] Run `make deploy` and smoketest during PR builds [skip ci]
Consider this a proof of concept. It uses DynamoDB conditional row updates to implement expiring mutexes. Most of that functionality is implememnted in a separate Python package at https://github.com/hannes-ucsc/lambda-mutex which is based on https://github.com/chiradeep/lambda-mutex. I add a fix for a race condition when the utility is run for the first time in a AWS region and added a CLI.

Reviewers should pay special attention to the Travis config.

The biggest problem with this is that master and PR builds would share a stage. A failed `make deploy` from a PR build could pollute the stage beyond repair, breaking subsequent builds. I think we should create a special sandbox environment for PR builds do decouple them from master.