Extract BundleDocument class from indexer
I would like to introduce a little OO into index.py. The benefit of this refactoring is that it makes more obvious how data flows between methods when preparing a document for indexing. The method signatures are simplified because contextual arguments are converted to instance attributes, for example `logger` and `replica`.

This change increases the line count mostly because it adds another way of constructing a BundleDocument from a ES document JSON.

This is a pure refactoring with minimal functional changes.

Reviewers are encouraged to review the diff with a tool that highlights intra-line changes, e.g. PyCharm. The diff will become very small as I paid close attention to not move code around, to the extent that one function became a method of BundleDocument even though it should not be. I will move those methods, deal with the overly long lines and remove the other TODOs once we reach consensus on whether we want this type of refactoring.