Rename 'bundle_id' to 'bundle_fqid'
The term 'bundle_id' has popped up in the repository, but it is not
always clear what it means.

Here I introduce the term Bundle Fully Qualified Identifier, which is
the combination of the bundle UUID and version string. It is analogous
to FQDN in that it specifies a bundle version exactly and fully.