Refactor bundle deletion test into reusable methods (connected to #816)
This splits the two cases (delete of a single version and of all versions) into separate test cases. It also reorders the creation of the tombstone to happen after the bundle is created. The original order would interfere with testing of a fix for #816. Having separate methods will make writing tests for #816 easier.