Run each test in a fresh set of ES indexes (connected to #918)
The previous approach of deleting documents and reusing the indexes for the next test proved to increase the probability of eventual consistency issues. The deleted documents would be gone from queries but remain in the index, raising a version conflict. This fix does not address the eventual consistency issue, it merely reduces its probability.

It also reduces index churm and increases test isolation. Keep in mind that the indexer tests reuse the same bundle version for mosts tests as a performance optimization.