[fix-s3-copy] support small files in s3 copy
Currently, the s3 copy step function flips out when it is passed a file that's too small for multipart copy.  It loops indefinitely in the first branch of the parallel section, which is not good.

This takes into account the possibility we might be asked to copy a small object.  If that happens, it copies during `setup_copy_task` and jumps straight to the end.