Fix: test_json_faker fails most of the time
Only every 16th build succeeds. The exact error message is

```
Error: Cannot create property '_randexp' on string '.{8}-.{4}-.{4}-.{4}-.{12}' in /properties/hca_ingest/properties/document_id
```

This is caused by a recent release of randexp, a transitive dependency, upgrading it from 0.4.6 to 0.4.7. The fix is to pin that dependency. In order to pin the dependency, we need to switch to using `package-lock.json` in conjuction with `npm install` which is the prescribed way of installing node.js dependencies reproducibly. Another fix would be to bump json-schema-faker to 0.4.7 but unfortunately that's not as easy because json-schema-faker-cli depends on it with a `^0.3.6` (essentially 0.3.x) spec. Bumping to 0.4.7 would conflict with that spec, causing json-schema-faker-cli to get its own node_modules directory, rendering the bump ineffective.