Remove line numbering from test output
It looks like sed buffers aggresively, exacerbating the risk of slow tests causing the 10min no-output timeout in Travis. Since we now run the really verbose, long-running and expensive tests in separate jobs the benefit of being able to correlate output to individual `make` processes and therefor test scripts is not as important anymore.