Remove backlink from file to bundle (fixes #1172)
This makes it possible for a file to belong to more than one bundle. It has implications on the physical deletion of files. We cannot remove a file without enumerating all bundles, loading their metadata and then checking if any other bundles reference that file. However this is no different to the physical deletion of blobs: since one blob can belong to many files, the same enumeration is required there. Ultimately, physical deletion due to lack of consent will always require the deletion of the actual data, i.e. the blob. These changes do not significantly complicate the search for files and bundles that refer to a blob, or, IOW, the presence of a backlink from file to bundle does not siginifcantly simplify the problem.