Fix duplicate notifications
Backend timeouts are currently not handled very well and could cause duplicate notifications. For the remainder of this text, note that we run the actual indexer backend as a task in a thread pool executor.

E = estimated time for writing one document (tombstone or bundle document)
R = remaining time in lambda execution

Currently the main thread times out after waiting E seconds for the backend thread and exits with an exception. The exception will eventually cause the lambda execution to fail and then to be retried. While the main thread is shutting down, the backend thread continues to run in its own pool thread. The pool isn't shut down (by choice, so it doesn't block) and the backend thread keeps running. It is a daemon however and therefore doesn't prevent the termination of the process, in this case the lambda execution context. In other words, it just gets terminated at a random point in its life. Just by chance, the runaway backend thread sometimes succeeds in queueing the notification for the indexed bundle. One of the two built-in retried execution of the lambda will then likely succeed and queue another duplicate notification.

Increasing E from 10s to 60s eliminated this problem but leaves the last 60s of a lambda execution unused. This is ok when indexing but not so much when reindexing where we want to index many bundles per lambda execution. We can do better by decoupling the time the main thread waits for the backend thread from E and use R instead. In other words, we'll leave in place the check that prevents the indexer from comitting to perform the indexing operations alltogether (update the index and queue the notification) if R < E but once it decided to do so, it will be given R instead of E to doing so.