Introduce get_bundle_manifest(..) (#1348)
storage/bundles transacts in what is stored on cloud storage.  The logic that translates what is stored in cloud storage into what the API endpoint is expecting is moved to dss.api.bundles.

This allows us to get rid of the awkward hack for compose_blob_key(..), as well as some unnecessary ['bundle'] indirection.