Convert existing s3copyclient code to use parallel_worker. (#1374)
Currently, we're marking the entire range (A, B) of a multipart copy complete if B completes. This is obviously not correct, as it only means B completes. (A, B-1) still may not have completed. Generally, this resolves itself as the parts do copy, but if the parts in (A, B-1) error out, then we don't ever complete the copy.

see dss-s3-copy-sfn-dev:9d5cb5f4-df3c-44ec-95bf-04dd54abebeb for an example of where this fails.

This rewrites the parallel copy management using the library in #1377.