Add envhook.py
This has proven useful in not having to start PyCharm or any other IDE
from a shell with `environment` sourced. Not only is that a nuissance,
you're essentially restricted to a single project because you can only
source the `environment` from a single project. If you have another
project that also uses an `environment`, you're out of luck. It
installs itself as a sitecustomize.py hook in the currently active
virtualenv. Whenever that virtualenv is activated, it will source
`environment` and apply the resulting environment changes to
`os.environ` such that the interpreter process and all its children
have see them.

To install, activate the virtualenv and run
`python scripts/envhook.py install`. To uninstall, run
`python scripts/envhook.py remove`.