SFN Sync daemon: bump Domovoi to build SNS-SQS bridge
When subscribing to S3 bucket events, only one subscriber is allowed
per event type. Subscribing an SQS queue and an SNS topic to the same
event type (dss-sync and dss-index) is impossible. This pulls in
changes to Domovoi that instead build an SNS-SQS bridge so that
the multiple subscribers are handled by SNS.

Fixup for 1cf7717