Add op for periodic replica consistency check
Adds a `scripts/dss-ops.py` target for use in periodically verifying the
consistency of given replicas, intended to be run on a schedule with Gitlab.
The time of last verification is stored in AWS Secrets Manager to allow for
"quick" consistency checks that only check objects that have not yet been
scanned by `sync verify-sync`.

The implemented operation, `sync verify-sync-all`, is a thin layer around
`sync verify-sync` that calls the latter with keys in the source replica
(so that those keys do not need to be provided manually).

This kind of verification does not need to be performed particularly often
and provides some assurance that sync is working as intended.

Related to #1680.