Back up collections, subscriptions (#2488)
Enables point-in-time recovery (PITR) for the
collections and subscripitons DynamoDB tables.
PITR allows an administrator to restore the
state of a table to its state at any time in the
last 35 days, which offers a degree of protection
against accidental data loss.

PITR is twice as expensive as creating on-demand
backups, but given the small size of the tables
that PITR is being enabled for, the cost is
negligible.