Update tombstone format for deletion RFC (#2516)
The deletion RFC specifies a tombstone comprised of keys `email`,
`reason`, and `details`. This commit removes the `version` key from
bundle tombstones and adds the `details` key. (It makes sense to
remove the version key anyway, since that information is already in
the tombstone name.)

Refactor tombstone helper function to dss.storage

We'll need to add support for tombstoning files, so it makes sense
to move the tombstone helper function to dss.storage from
dss.api.bundles.